+++ zcash_client_backend/src/data_api/testing/pool.rs
@@ -3349,6 +3349,7 @@ where
         spent_outpoint.clone(),
         TxOut::new(Zatoshis::const_from_u64(100000), taddr.script().into()),
         Some(h),
+        Some(TransparentKeyScope::EXTERNAL),
     )
     .unwrap();
 
@@ -6050,8 +6051,13 @@ pub fn wallet_recovery_computes_fees<T: ShieldedPoolTester, DsF: DataStoreFactor
         assert_eq!(t_bundle.vout.len(), 1);
 
         let outpoint = OutPoint::new(*txid.as_ref(), 0);
-        let utxo = WalletTransparentOutput::from_parts(outpoint, t_bundle.vout[0].clone(), Some(h))
-            .unwrap();
+        let utxo = WalletTransparentOutput::from_parts(
+            outpoint,
+            t_bundle.vout[0].clone(),
+            Some(h),
+            Some(TransparentKeyScope::EXTERNAL),
+        )
+        .unwrap();
         st.wallet_mut()
             .put_received_transparent_utxo(&utxo)
             .unwrap();
@@ -6343,6 +6349,7 @@ where
         OutPoint::fake(),
         TxOut::new(non_coinbase_value, t_addr.script().into()),
         Some(h),
+        Some(TransparentKeyScope::EXTERNAL),
     )
     .unwrap();
     st.wallet_mut()