+++ zcash_client_sqlite/src/wallet.rs
@@ -3433,6 +3433,7 @@ pub(crate) fn store_transaction_to_be_sent<P: consensus::Parameters>(
                                     ),
                                     TxOut::new(output.value(), taddr.script().into()),
                                     None,
+                                    Some(TransparentKeyScope::EXTERNAL)
                                 )
                                 .expect(
                                     "can extract a recipient address from an internal address script",
@@ -3505,6 +3506,7 @@ pub(crate) fn store_transaction_to_be_sent<P: consensus::Parameters>(
                         outpoint.clone(),
                         TxOut::new(output.value(), ephemeral_address.script().into()),
                         None,
+                        Some(TransparentKeyScope::EPHEMERAL),
                     )
                     .expect("can extract a recipient address from an ephemeral address script"),
                     sent_tx.target_height().into(),